Move advice crawl and helper resolution to a separate HelperResolver
